Обновил префабы. Перенес функциональность на сцену RockyIsland. Добавил комментарии. Занес рефакторинг взаимодейтсивя/тултипов в тех. долг.
