+++ Assets/Game/UI/InteractableObjectTooltip.prefab.meta
