--- Assets/Scripts/IInteractible.cs
@@ -1,10 +1,0 @@
-﻿using UnityEngine;
-
-public interface IInteractable : ITooltipable
-{
-    public bool IsInteractionAllowed();
-    
-    public void Interact();
-
-    public float GetDistanceTo(Transform otherTransform);
-}
