--- Assets/Scripts/ITooltipable.cs
@@ -1,11 +1,0 @@
-﻿using UnityEngine;
-
-public interface ITooltipable
-{
-    
-    public string GetTooltipTitle();
-
-    public string GetTooltipDescription();
-    
-    public Vector3 GetWorldPosition();
-}
