+++ Assets/Scripts/InputSystem/InputButton.cs
@@ -32,7 +32,8 @@
 
         /// <summary> Открыть вкладку настроек в игровом меню. </summary>
         OpenSettingsTab,
-        
+
+        /// <summary> Взаимодействие с объектами. </summary>
         Interact,
 
         /// <summary> Позиция курсора мыши. </summary>
