+++ Assets/Scripts/TooltipSystem/InteractableObjectToolTip.cs.meta
